@@ -3,7 +3,7 @@ source 'https://rubygems.org'
 gem 'dotenv-rails'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.10'
+gem 'rails', '5.0.7'
 
 gem 'pg', '~> 0.15'
 
@@ -99,6 +99,7 @@ group :test do
   gem 'climate_control'
   gem "vcr"
   gem "webmock"
+  gem 'rails-controller-testing'
 end
 
 group :development, :test do
