@@ -1,6 +1,6 @@
 class TrackingController < ApplicationController
   # We need this controller to be accessible by anyone without authentication
-  skip_filter :authenticate_admin!
+  skip_before_action :authenticate_admin!
   # SSL redirection is also disabled for this controller. See ApplicationController force_ssl.
 
   def open2
