@@ -12,7 +12,7 @@ describe AppsController, type: :controller do
       before(:each) { sign_in admin }
 
       it "should create an app that is part of the current user's team" do
-        post :create, app: {name: "My New App"}
+        post :create, params: { app: {name: "My New App"} }
         app = App.where(cuttlefish: false).first
         expect(app.name).to eq "My New App"
         expect(app.team).to eq team
