@@ -15,7 +15,7 @@ describe InvitationsController, type: :controller do
 
       it "should invite a user by their email address and make them part of the team" do
         expect(Admin).to receive(:invite!).with({"email" => "matthew@foo.bar", "team_id" => team.id}, admin).and_call_original
-        post :create, admin: {email: "matthew@foo.bar"}
+        post :create, params: { admin: {email: "matthew@foo.bar"} }
         expect(response).to redirect_to admins_url
       end
     end
